fix (converter): avoid circular dep.
