+++ src/models/v2/lists/list-node.ts
@@ -1,7 +1,45 @@
-import { JsonObject, JsonProperty } from "json2typescript";
+import {
+    JsonConvert,
+    JsonConverter,
+    JsonCustomConvert,
+    JsonObject,
+    JsonProperty,
+    OperationMode,
+    ValueCheckingMode
+} from "json2typescript";
+import { PropertyMatchingRule } from "json2typescript/src/json2typescript/json-convert-enums";
 import { Constants } from "../Constants";
 import { IdConverter } from "../custom-converters/id-converter";
-import { SubListNodeConverter } from "../custom-converters/sublist-node-converter";
+
+@JsonConverter
+export class SubListNodeConverter implements JsonCustomConvert<ListNode[]> {
+
+    static jsonConvert: JsonConvert = new JsonConvert(
+        OperationMode.ENABLE,
+        ValueCheckingMode.DISALLOW_NULL,
+        false,
+        PropertyMatchingRule.CASE_STRICT
+    );
+
+    serialize(subclasses: ListNode[]): any {
+    }
+
+    deserialize(subnodes: any): ListNode[] {
+
+        let children: object[];
+
+        if (Array.isArray(subnodes)) {
+            children = subnodes;
+        } else {
+            children = [subnodes];
+        }
+
+        return children
+            .map(
+                child =>
+                    SubListNodeConverter.jsonConvert.deserialize(child, ListNode) as ListNode);
+    }
+}
 
 @JsonObject("ListNode")
 export class ListNode {
